pyproject.toml
[tool.poetry]
name = "stream-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Sustavov Maksim <sustavovm@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.10"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.13.3"
celery = {extras = ["redis"], version = "^5.4.0"}
requests = "^2.32.3"
selenium = "^4.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.6"
pytest-vcr = "^1.0.2"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"
pytest-dotenv = "^0.5.2"
pytest-celery = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
